sprint5/demo5/requirements.txt
alembic==1.7.7
astroid==2.11.2
black==22.3.0
click==8.1.2
dill==0.3.4
Flask==2.1.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.1
lazy-object-proxy==1.7.1
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
pylint==2.13.4
python-dotenv==0.20.0
SQLAlchemy==1.4.34
tomli==2.0.1
typing_extensions==4.1.1
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.8.0
